+++ src/components/elements/Copyright/Copyright.tsx
@@ -5,9 +5,9 @@ import React, { ReactElement } from 'react'
 export default function Copyright(): ReactElement {
     return (
         <Typography variant="body2" color="textSecondary" align="center">
-            {'No Copyright © '} 
-            <Link color="inherit" href="https://martinreboredo.com/">
-                Martin Reboredo
+            {'Copyright © '} 
+            <Link color="inherit" href="https://www.netenders.com/">
+                Netenders Corp. 
                 </Link>
             <Box component="span" display="inline" data-testid="date" ml={1}>
                 {new Date().getFullYear()}