+++ src/components/elements/Hero/Hero.tsx
@@ -26,21 +26,16 @@ export default function Album() {
         <div className={classes.heroContent}>
           <Container maxWidth="sm">
             <Typography component="h1" variant="h2" align="center" color="textPrimary" gutterBottom>
-              MERN
+              (NAP)
             </Typography>
             <Typography variant="h5" align="center" color="textSecondary" paragraph>
-              Boilerplate Code for a MERN. Using Docker, Redux, Google Cloud, Material-UI.
+              Netenders Administration Panel 
             </Typography>
             <div className={classes.heroButtons}>
               <Grid container spacing={2} justify="center">
                 <Grid item>
-                  <Button href="https://github.com/martinrebo/boilerplate" target="_blank" variant="contained" color="primary">
-                    See Source Code
-                  </Button>
-                </Grid>
-                <Grid item>
-                  <Button href="https://www.martinreboredo.com/" target="_blank" variant="outlined" color="primary">
-                    Contact Me
+                  <Button href="/" target="_blank" variant="contained" color="primary">
+                    Login
                   </Button>
                 </Grid>
               </Grid>